requirements.txt
snowflake-connector-python==2.27